test(abstract-eth): move address init tests to abstract-eth WIN-1926
This PR moves address init tests to abstract-eth
TICKET: WIN-1926
