+++ modules/abstract-eth/test/unit/transactionBuilder/index.ts
@@ -1,0 +1,1 @@
+export * from './addressInitialization';
