+++ modules/sdk-coin-arbeth/test/resources.ts
@@ -79,6 +79,10 @@ export const BASE_ADDRESS = '0xbc42187c73d2f41a5551be6d138298f4e9f2e2c6';
 
 export const FEE_ADDRESS = '0x5fc4bc0f18c4243ee4ce757d712bb2164176cb51';
 
+export const DEPLOYED_ADDRESS = '0x035540896df72a28bff04f6800048ade739e6d05';
+
+export const SENDER_ADDRESS = '0xe6c43626f11312de29b0011fa9da71ea3bba0e9f';
+
 export const UNSIGNED_ADDRESS_INIT =
   '0xf887010a8203e894e0fce739287037751555011785c369fb1bdfe9c880b86413b2f75c000000000000000000000000bc42187c73d2f41a5551be6d138298f4e9f2e2c600000000000000000000000000000000000000000000000000000000000000010000000000000000000000005fc4bc0f18c4243ee4ce757d712bb2164176cb51830cddff8080';
 
@@ -87,3 +91,9 @@ export const SIGNED_ADDRESS_INIT =
 
 export const TX_BROADCAST =
   '0xf9012d018502540be4008367c28094252a69d9a2be99aae36a80ed92d311cf38ec89ce80b8c47117f3fa000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003000000000000000000000000e6c43626f11312de29b0011fa9da71ea3bba0e9f00000000000000000000000078caeb4527170e52f54d936e4eef6f83250e01bb000000000000000000000000b1938215967408fff7c59c77ae5e5283b55c8e26830cddffa0c2c50b071badc15d4009a77c9194df6938b0148a3c04792c60d793132023be58a04e20e1a1c9bfee5099feebec6bcfc6d9c9bf0ae3ae175c4d53f1ac52bdb31864';
+
+export const SIGNED_INIT_TX_ID = '0xdc7ebdb8aa3c190888825fb38a29ff0530247a436df0204577b95de96bf6baef';
+
+export const CHAIN_ID = '0x66eee';
+
+export const FINAL_V = '0x0cddff';
