+++ modules/sdk-coin-arbeth/test/unit/transactionBuilder/addressInitialization.ts
@@ -1,87 +1,8 @@
-import should from 'should';
-import { TransactionType } from '@bitgo/sdk-core';
+import { runAddressInitializationTests } from '@bitgo/abstract-eth';
 import { TransactionBuilder } from '../../../src';
 import * as testData from '../../resources';
 import { getBuilder } from '../../getBuilder';
 
-describe('Arbeth address initialization', () => {
-  it('should fail if there is no contract address', async () => {
-    const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
-    txBuilder.type(TransactionType.AddressInitialization);
-    txBuilder.fee({
-      fee: '10',
-      gasLimit: '1000',
-    });
-    txBuilder.counter(1);
-    await txBuilder.build().should.be.rejectedWith('Invalid transaction: missing contract address');
-  });
+const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
 
-  it('should fail if forwarder version is invalid', async () => {
-    const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
-    try {
-      txBuilder.type(TransactionType.AddressInitialization);
-      txBuilder.fee({
-        fee: '10',
-        gasLimit: '1000',
-      });
-      txBuilder.contract(testData.FORWARDER_FACTORY_ADDRESS);
-      txBuilder.forwarderVersion(3);
-    } catch (e) {
-      should.equal(e.message, 'Invalid forwarder version: 3');
-    }
-  });
-
-  it('should build properly and return a correct address', async () => {
-    const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
-    txBuilder.type(TransactionType.AddressInitialization);
-    txBuilder.fee({
-      fee: '10',
-      gasLimit: '1000',
-    });
-    txBuilder.counter(1);
-    txBuilder.contract(testData.FORWARDER_FACTORY_ADDRESS);
-    txBuilder.baseAddress(testData.BASE_ADDRESS);
-    txBuilder.feeAddress(testData.FEE_ADDRESS);
-    txBuilder.forwarderVersion(4);
-    txBuilder.salt('0x1');
-    txBuilder.initCode(testData.FORWARDER_IMPLEMENTATION_ADDRESS);
-    txBuilder.sign({ key: testData.KEYPAIR_PRV.getKeys().prv });
-    const tx = await txBuilder.build();
-    const txJson = tx.toJson();
-    should.equal(txJson.deployedAddress, '0x035540896df72a28bff04f6800048ade739e6d05');
-    should.equal(txJson.to, testData.FORWARDER_FACTORY_ADDRESS);
-  });
-
-  it('should build properly from serialized', async () => {
-    const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
-    txBuilder.type(TransactionType.AddressInitialization);
-    txBuilder.from(testData.UNSIGNED_ADDRESS_INIT);
-    const tx = await txBuilder.build();
-    const txJson = tx.toJson();
-    should.equal(txJson.to, testData.FORWARDER_FACTORY_ADDRESS);
-  });
-
-  it('should build properly from signed serialized', async () => {
-    const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
-    txBuilder.type(TransactionType.AddressInitialization);
-    txBuilder.from(testData.SIGNED_ADDRESS_INIT);
-    const tx = await txBuilder.build();
-    const txJson = tx.toJson();
-    should.equal(txJson.to, testData.FORWARDER_FACTORY_ADDRESS);
-    should.equal(txJson.from, '0xe6c43626f11312de29b0011fa9da71ea3bba0e9f');
-  });
-
-  it('should build properly createForwarder call for recovery', async () => {
-    const txBuilder = getBuilder('tarbeth') as TransactionBuilder;
-    txBuilder.type(TransactionType.ContractCall);
-    txBuilder.contract(testData.BASE_ADDRESS);
-    txBuilder.data(testData.CREATE_FORWARDER_METHOD);
-    txBuilder.fee({
-      fee: '10',
-      gasLimit: '1000',
-    });
-    const tx = await txBuilder.build();
-    const txJson = tx.toJson();
-    should.equal(txJson.to, testData.BASE_ADDRESS);
-  });
-});
+runAddressInitializationTests('tarbeth', txBuilder, testData);
