+++ modules/sdk-coin-opeth/test/resources.ts
@@ -79,6 +79,10 @@ export const BASE_ADDRESS = '0xbc42187c73d2f41a5551be6d138298f4e9f2e2c6';
 
 export const FEE_ADDRESS = '0x5fc4bc0f18c4243ee4ce757d712bb2164176cb51';
 
+export const DEPLOYED_ADDRESS = '0xab45abb1dcd3eebc37cac952f753fff687e2a34a';
+
+export const SENDER_ADDRESS = '0xe6c43626f11312de29b0011fa9da71ea3bba0e9f';
+
 export const UNSIGNED_ADDRESS_INIT =
   '0xf888010a8203e894ce54336e87a5cb5440d38c32d996c3ecb3a5968f80b86413b2f75c000000000000000000000000bc42187c73d2f41a5551be6d138298f4e9f2e2c600000000000000000000000000000000000000000000000000000000000000010000000000000000000000005fc4bc0f18c4243ee4ce757d712bb2164176cb518401546fdb8080';
